fullstack/backend/python/requirements.txt
flask
flask-cors 
typedb-driver
